sycl: move sycl_compat.hpp to util subdir
