+++ src/umpire/op/SyclMemPrefetchOperation.cpp
@@ -6,7 +6,7 @@
 //////////////////////////////////////////////////////////////////////////////
 #include "umpire/op/SyclMemPrefetchOperation.hpp"
 
-#include "umpire/sycl_compat.hpp"
+#include "umpire/util/sycl_compat.hpp"
 
 #include "umpire/strategy/AllocationStrategy.hpp"
 #include "umpire/util/error.hpp"
