+++ src/umpire/resource/MemoryResourceRegistry.cpp
@@ -50,7 +50,7 @@
 #endif
 
 #if defined(UMPIRE_ENABLE_SYCL)
-#include "umpire/sycl_compat.hpp"
+#include "umpire/util/sycl_compat.hpp"
 
 #include "umpire/resource/SyclDeviceResourceFactory.hpp"
 #include "umpire/resource/SyclPinnedMemoryResourceFactory.hpp"
